extract BottomNavBarWidget into seperate file for code modularity
