+++ lib/ui_components/bottom_nav_bar.dart
@@ -1,0 +1,66 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+/// Bottom navigation bar widget with an app bar and tab icons.
+class BottomNavBar extends StatelessWidget {
+  const BottomNavBar({
+    required this.child,
+    super.key,
+  });
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: child,
+      bottomNavigationBar: BottomNavigationBar(
+        items: const [
+          BottomNavigationBarItem(
+            icon: Icon(Icons.home),
+            label: 'Home',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.search),
+            label: 'Detections',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.show_chart),
+            label: 'Stats',
+          ),
+        ],
+        currentIndex: _calculateSelectedIndex(context),
+        onTap: (int idx) => _onItemTapped(idx, context),
+      ),
+    );
+  }
+
+  // Calculate the index of the bottom navigation bar based on the current route
+  static int _calculateSelectedIndex(BuildContext context) {
+    final String location = GoRouterState.of(context).uri.toString();
+    if (location == '/main') {
+      return 0;
+    }
+    if (location == '/main/detections') {
+      return 1;
+    }
+    if (location == '/main/stats') {
+      return 2;
+    }
+    return 0;
+  }
+
+  // Function to handle navigation when an item is tapped
+  void _onItemTapped(int index, BuildContext context) {
+    switch (index) {
+      case 0:
+        GoRouter.of(context).go('/main');
+        break;
+      case 1:
+        GoRouter.of(context).go('/main/detections');
+        break;
+      case 2:
+        GoRouter.of(context).go('/main/stats');
+    }
+  }
+}
